+++ app/src/main/java/nextstep/signup/ui/signup/SignUpScreen.kt
@@ -1,0 +1,120 @@
+package nextstep.signup.ui.signup
+
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material3.Button
+import androidx.compose.material3.ButtonDefaults
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.setValue
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.res.colorResource
+import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.text.style.TextAlign
+import androidx.compose.ui.tooling.preview.Preview
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+import nextstep.signup.R
+import nextstep.signup.ui.component.EmailInput
+import nextstep.signup.ui.component.PasswordConfirmInput
+import nextstep.signup.ui.component.PasswordInput
+import nextstep.signup.ui.component.UserNameInput
+import nextstep.signup.ui.theme.SignupTheme
+
+@Composable
+fun SignUpScreen(
+    modifier: Modifier = Modifier
+) {
+    var userName by remember { mutableStateOf("") }
+    var email by remember { mutableStateOf("") }
+    var password by remember { mutableStateOf("") }
+    var passwordConfirm by remember { mutableStateOf("") }
+
+    Column(
+        modifier = Modifier
+            .fillMaxSize()
+            .padding(20.dp),
+        horizontalAlignment = Alignment.CenterHorizontally
+    ) {
+        Text(
+            modifier = Modifier
+                .fillMaxWidth()
+                .padding(top = 40.dp),
+            text = stringResource(id = R.string.title),
+            fontSize = 26.sp,
+            fontWeight = FontWeight.Bold,
+            textAlign = TextAlign.Center
+        )
+
+        UserNameInput(
+            value = userName,
+            onValueChange = {
+                userName = it
+            },
+            modifier = modifier
+        )
+
+        EmailInput(
+            value = email,
+            onValueChange = {
+                email = it
+            },
+            modifier = modifier
+        )
+
+        PasswordInput(
+            value = password,
+            onValueChange = {
+                password = it
+            },
+            modifier = modifier
+        )
+
+        PasswordConfirmInput(
+            value = passwordConfirm,
+            password = password,
+            onValueChange = {
+                passwordConfirm = it
+            },
+            modifier = modifier
+        )
+
+        Button(
+            onClick = { /*TODO*/ },
+            colors = ButtonDefaults.buttonColors(
+                disabledContainerColor = colorResource(R.color.blue),
+                disabledContentColor = Color.White
+            ),
+            content = {
+                Text(
+                    text = stringResource(id = R.string.signUp)
+                )
+            },
+            modifier = Modifier
+                .fillMaxWidth()
+                .height(100.dp)
+                .padding(20.dp)
+        )
+    }
+}
+
+@Preview
+@Composable
+private fun SignUpScreenPreview() {
+    SignupTheme {
+        SignUpScreen(
+            modifier = Modifier
+                .fillMaxWidth()
+                .padding(top = 20.dp)
+        )
+    }
+}
