+++ app/src/main/res/values/strings.xml
@@ -9,7 +9,6 @@
     <string name="userNameInvalidCharactersMessage">이름에는 숫자나 기호가 포함될 수 없습니다.</string>
     <string name="userNameLengthValidationMessage">이름은 2~5자여야 합니다.</string>
     <string name="emailInvalidFormatMessage">이메일 형식이 올바르지 않습니다.</string>
-    <string name="passwordLengthValidationMessage">비밀번호는 8~16자여야 합니다.</string>
-    <string name="passwordComplexityMessage">비밀번호는 영문과 숫자를 포함해야 합니다.</string>
+    <string name="passwordLValidationMessage">비밀번호는 8~16자여야 합니다.\n비밀번호는 영문과 숫자를 포함해야 합니다</string>
     <string name="passwordMismatchMessage">비밀번호가 일치하지 않습니다.</string>
 </resources>
